+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - explanation about forks and token permissions with them.
+- new option `skip-comments-on-forks` to prevent blocking pull requests from forks because of lack of
+  permissions.
 
 ### Changed